Add "Add Book" Functionality
Added the "add book" functionality. Users are now able to click a button that displays a modal that contains a form. Inside the form, the users are able to input the book title and book author. More input options will definitely be added in the future. The modal can also be closed using the 'x' button in the right corner. The modal also closes when clicking the 'add book'/'submit' button. The information used in the form is used to create a new book and add it to the DOM/display.